+++ src/tests/integration/extension.test.ts
@@ -1,6 +1,6 @@
 // gitgerbil-ignore-file
 
-import { defaultScannedFiles } from "../../extension";
+import { defaultScannedFiles } from "../../extension/utils/globals";
 import { describe } from "mocha";
 import assert from "node:assert";
 import * as vscode from "vscode";