+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Updated the README to include `enableStrictSecretScanning` in the list of settings.
+- Tweaked regex patterns for all 3 types of scanning to reduce false positives.
+- Reduced package size significantly by removing unnecessary `hero.png` asset from package.
+
 ## [0.1.4] - 2026-02-10
 
 ### Added