+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.6] - 2026-02-12
+
+### Changed
+
+- Fixed an issue where quick fix code actions were being provided for external diagnostics.
+
 ## [0.1.5] - 2026-02-11
 
 ### Added